--- css/splash.css
@@ -1,10 +1,0 @@
-.banner h1 {
-    display: block;
-    font-size: 4.8rem;
-    letter-spacing: 0;
-    line-height: 0.8;
-    padding: 1em 0 0;
-    text-transform: uppercase;
-    font-weight: 900;
-    color: #182e4f;
-}
