+++ frontend/src/components/main/main_page.js
@@ -1,18 +1,12 @@
 import React from 'react';
+import MainSearchComponent from './main_seach_component';
 
 class MainPage extends React.Component {
 
     render() {
         return (
             <div>
-                <section className="intro">
-                    <header className="banner">
-                        <h1>About Time</h1>
-                    </header>
-                </section>
-            <footer>
-                Watch recommendations
-            </footer>
+                <MainSearchComponent />
             </div>
         );
     }
