+++ frontend/src/css/splash.css
@@ -1,0 +1,96 @@
+.banner h1 {
+    display: block;
+    font-size: 4.8rem;
+    letter-spacing: 0;
+    line-height: 0.8;
+    padding: 1em 0 0;
+    text-transform: uppercase;
+    font-weight: 900;
+    color: #182e4f;
+}
+
+/*Search*/
+body {
+    background-color: #fff5d7;
+}
+.splash-form {
+    background-color: #ffffff;
+    margin: 100px auto;
+    padding: 40px;
+    width: 70%;
+    max-width: 60%;
+    min-width: 300px;
+    border-radius: 15px;
+    -webkit-box-shadow: 0 10px 6px -6px #777;
+    -moz-box-shadow: 0 10px 6px -6px #777;
+    box-shadow: 0 10px 6px -6px #777;
+}
+
+.splash-form h1 {
+    text-align: center;  
+    font-family: 'Lato', sans-serif;
+}
+
+.splash-form input{
+  padding: 10px;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  width: 100%;
+  font-size: 17px;
+  font-family: Raleway;
+  border: 1px solid #aaaaaa;
+}
+
+.splash-form label {
+    font-size: 17px;
+    font-family: Raleway;
+}
+.splash-form select {
+	display: block;
+	font-size: 16px;
+	font-family: sans-serif;
+	font-weight: 700;
+	color: #444;
+	line-height: 1.3;
+	padding: .6em 1.4em .5em .8em;
+	width: 100%;
+	max-width: 100%;
+	box-sizing: border-box;
+	margin: 0;
+	border: 1px solid #aaa;
+	box-shadow: 0 1px 0 1px rgba(0,0,0,.04);
+    border-radius: .5em;
+    margin-top: 20px;
+}
+
+.splash-form option {
+    font-family: sans-serif;
+    color: #444;
+}
+
+.splash-form-submit {
+    box-shadow:inset 0px 1px 0px 0px #ffffff;
+	background:linear-gradient(to bottom, #ededed 5%, #dfdfdf 100%);
+	background-color:#ededed;
+	border-radius:6px;
+	border:1px solid #dcdcdc;
+	display:inline-block;
+	cursor:pointer;
+	color:#777777;
+	font-family:Arial;
+	font-size:12px;
+	font-weight:bold;
+	padding:4px 18px;
+	text-decoration:none;
+	text-shadow:0px 1px 0px #ffffff;
+}
+
+.splash-form-submit:hover {
+    background:linear-gradient(to bottom, #dfdfdf 5%, #ededed 100%);
+	background-color:#dfdfdf;
+}
+
+.splash-form-submit:active {
+    position:relative;
+	top:1px;
+}
