+++ frontend/src/util/watch_api_util.js
@@ -1,3 +1,4 @@
+import axios from 'axios';
 export const getWatch = (id) => {
     return axios.get(`/api/watches/${id}`)
 }
